# ARROW-8494: [C++][Parquet] Full support for reading mixed list and structs