--- a/cpp/src/parquet/arrow/reader_internal.h
+++ b/cpp/src/parquet/arrow/reader_internal.h
@@ -104,13 +104,6 @@ Status TransferColumnData(::parquet::internal::RecordReader* reader,
                           const ColumnDescriptor* descr, ::arrow::MemoryPool* pool,
                           std::shared_ptr<::arrow::ChunkedArray>* out);
 
-Status ReconstructNestedList(const std::shared_ptr<::arrow::Array>& arr,
-                             std::shared_ptr<::arrow::Field> field, int16_t max_def_level,
-                             int16_t max_rep_level, const int16_t* def_levels,
-                             const int16_t* rep_levels, int64_t total_levels,
-                             ::arrow::MemoryPool* pool,
-                             std::shared_ptr<::arrow::Array>* out);
-
 struct ReaderContext {
   ParquetFileReader* reader;
   ::arrow::MemoryPool* pool;