… cpp/src/parquet/CMakeLists.txt
… cpp/src/parquet/arrow/arrow_reader_writer_test.cc
… cpp/src/parquet/arrow/reader_internal.cc
… cpp/src/parquet/arrow/reconstruct_internal_test.cc
… cpp/src/parquet/column_reader.cc
… cpp/src/parquet/column_reader.h
… cpp/src/parquet/level_comparison.cc
… cpp/src/parquet/level_comparison.h
… cpp/src/parquet/level_comparison_avx2.cc
… cpp/src/parquet/level_comparison_inc.h
… cpp/src/parquet/level_conversion.cc
… cpp/src/parquet/level_conversion.h
… cpp/src/parquet/level_conversion_benchmark.cc
… cpp/src/parquet/level_conversion_bmi2.cc
… cpp/src/parquet/level_conversion_inc.h
… python/pyarrow/tests/test_parquet.py