@pitrou 2020-09-14
Are there any benchmarks worth running here?