@emkornfield 2020-09-15
> Ok, thanks a lot for this PR. I think I am understanding the implementation (I skipped parquet/arrow/reader.cc for now, though). Some of the implementation details are still confusing me a bit. In any case, here are some comments.

Please let me know if there is more confusion, I will attempt to add clarifying comments.  I think I addressed all your comments except for some in level_conversion_test.cc I'll address those tomorrow (I assume there will be more comments in reader.cc as well).

> Are there any benchmarks worth running here?

parquet-level-conversion-benchmark
parquet-arrow-reader-writer-benchmark (this won't cover the nested cases though)  There is an open JIRA under ARROW-1644 to add benchmarks for nested cases @npr mentioned there might be some example datasets that we wanted to try this on.