@pitrou 2020-09-15
`parquet-level-conversion-benchmark`:
* before:
```
BM_DefinitionLevelsToBitmapRepeatedAllMissing         966 ns          966 ns       714328 bytes_per_second=1.9753G/s
BM_DefinitionLevelsToBitmapRepeatedAllPresent        1626 ns         1626 ns       426358 bytes_per_second=1.173G/s
BM_DefinitionLevelsToBitmapRepeatedMostPresent       1850 ns         1850 ns       367809 bytes_per_second=1055.83M/s
```
* after:
```
BM_DefinitionLevelsToBitmapRepeatedAllMissing         560 ns          560 ns      1239730 bytes_per_second=3.40515G/s
BM_DefinitionLevelsToBitmapRepeatedAllPresent         562 ns          562 ns      1244256 bytes_per_second=3.39684G/s
BM_DefinitionLevelsToBitmapRepeatedMostPresent        562 ns          562 ns      1244691 bytes_per_second=3.39515G/s
```

Impressive!