@pitrou 2020-09-15
No changes on `parquet-arrow-reader-writer-benchmark`. I suspect it doesn't trigger any of the updated code?