@emkornfield 2020-09-16
@pitrou I think I responded to all review comments at this point, apologies if I missed something.  level_conversion_test.cc is now refactored to a point where duplicate code I think adds to test understandability but there is still some redundancy.  Also, please see my note about the level_conversion_benchmark having a bug in it on your prior run.