@nealrichardson 2020-09-16
I think the macOS failure is fixed by #8196 but the Appveyor failure looks legit:

```
C:/projects/arrow/cpp/src/parquet/level_conversion_test.cc(295): error C2220: warning treated as error - no 'object' file generated
C:/projects/arrow/cpp/src/parquet/level_conversion_test.cc(278): note: while compiling class template member function 'void parquet::internal::NestedListTest_SimpleLongList_Test<T>::TestBody(void)'
        with
        [
            T=Type
        ]
googletest_ep-prefix\include\gtest/internal/gtest-internal.h(665): note: see reference to class template instantiation 'parquet::internal::NestedListTest_SimpleLongList_Test<T>' being compiled
        with
        [
            T=Type
        ]
googletest_ep-prefix\include\gtest/internal/gtest-internal.h(657): note: while compiling class template member function 'bool testing::internal::TypeParameterizedTest<parquet::internal::NestedListTest,testing::internal::TemplateSel<parquet::internal::NestedListTest_SimpleLongList_Test>,parquet::internal::gtest_type_params_NestedListTest_>::Register(const char *,const testing::internal::CodeLocation &,const char *,const char *,int,const std::vector<std::string,std::allocator<_Ty>> &)'
        with
        [
            _Ty=std::string
        ]
C:/projects/arrow/cpp/src/parquet/level_conversion_test.cc(278): note: see reference to function template instantiation 'bool testing::internal::TypeParameterizedTest<parquet::internal::NestedListTest,testing::internal::TemplateSel<parquet::internal::NestedListTest_SimpleLongList_Test>,parquet::internal::gtest_type_params_NestedListTest_>::Register(const char *,const testing::internal::CodeLocation &,const char *,const char *,int,const std::vector<std::string,std::allocator<_Ty>> &)' being compiled
        with
        [
            _Ty=std::string
        ]
C:/projects/arrow/cpp/src/parquet/level_conversion_test.cc(278): note: see reference to class template instantiation 'testing::internal::TypeParameterizedTest<parquet::internal::NestedListTest,testing::internal::TemplateSel<parquet::internal::NestedListTest_SimpleLongList_Test>,parquet::internal::gtest_type_params_NestedListTest_>' being compiled
C:/projects/arrow/cpp/src/parquet/level_conversion_test.cc(295): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
[212/245] Linking CXX shared library release\parquet.dll
   Creating library release\parquet.lib and object release\parquet.exp
[213/245] Building CXX object src\parquet\CMakeFiles\parquet-arrow-test.dir\Unity\unity_0_cxx.cxx.obj
ninja: build stopped: subcommand failed.
```