@pitrou 2020-09-17
I'll take a look again on Monday, if that's ok with you.