@pitrou 2020-09-21
Running `parquet-level-conversion-benchmark` again, results are much less good:
```
BM_DefinitionLevelsToBitmapRepeatedAllMissing         878 ns          878 ns       816435 bytes_per_second=2.17265G/s
BM_DefinitionLevelsToBitmapRepeatedAllPresent         967 ns          966 ns       713185 bytes_per_second=1.9736G/s
BM_DefinitionLevelsToBitmapRepeatedMostPresent       3258 ns         3258 ns       214817 bytes_per_second=599.563M/s
```