@pitrou 2020-09-21
Things are a bit more balanced if the scalar version is used:
```
BM_DefinitionLevelsToBitmapRepeatedAllMissing         660 ns          660 ns      1019688 bytes_per_second=2.8884G/s
BM_DefinitionLevelsToBitmapRepeatedAllPresent        1946 ns         1946 ns       358335 bytes_per_second=1003.9M/s
BM_DefinitionLevelsToBitmapRepeatedMostPresent       2024 ns         2023 ns       346078 bytes_per_second=965.229M/s
```