@pitrou 2020-09-21
Just for the record, apart from FixedSizeList, is there anything remaining for full nested Parquet -> Arrow reading?