@pitrou 2020-09-21
Other than that, I see a ~20% improvement on `BM_ReadStructColumn` and `BM_ReadListColumn`