@pitrou 2020-09-21
I have no remaining concern over the code other than the AVX2 / BMI2 split. Congratulations for this PR, this is really a huge improvement!

That said, I seem to get a test error on the Python side (pasted below). Let's see if it reproduces on CI:
```
Traceback (most recent call last):
  File "/home/antoine/arrow/dev/python/pyarrow/tests/test_parquet.py", line 700, in test_pandas_can_write_nested_data
    _write_table(arrow_table, imos)
  File "/home/antoine/miniconda3/envs/pyarrow/lib/python3.7/site-packages/_pytest/python_api.py", line 747, in __exit__
    fail(self.message)
  File "/home/antoine/miniconda3/envs/pyarrow/lib/python3.7/site-packages/_pytest/outcomes.py", line 128, in fail
    raise Failed(msg=msg, pytrace=pytrace)
Failed: DID NOT RAISE <class 'ValueError'>
```

(beware: I rebased your branch on git master)