@emkornfield 2020-09-21
> I have no remaining concern over the code other than the AVX2 / BMI2 split. Congratulations for this PR, this is really a huge improvement!

@pitrou thank you for the thoughtful review.  Let me know if you still have issues with the AVX/BMI2 after I added the comment (perhaps I didn't revert some compilation) or my analysis is wrong.  I think the BMI2 check will be difficult/impossible at compile time for windows, so I'm not sure if it is worth the effort  on linux.

I also removed the failing test for parquet (which I should have removed in a prior PR, its strange it showed up again).